pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "maple"
version = "0.1.0"
description = "LLM workflow automation tools for extracting QSP metadata from scientific literature"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Joel Eliason", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "openai>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "pandas>=2.0.0",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "pydantic-ai>=0.0.14",  # For structured outputs with discriminated unions
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "pint>=0.25",  # For unit-aware test statistic calculations
    "ruamel.yaml>=0.18",  # Better YAML formatting with block scalars
    "scipy>=1.11.0",  # For ODE integration in IsolatedSystemTarget validation
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black==24.10.0",
    "ruff>=0.1.0",
    "freezegun>=1.2.0",
]

[project.scripts]
# Main workflow commands
qsp-extract = "maple.cli.extract:main"
qsp-validate = "maple.cli.validate:main"
qsp-review = "maple.cli.review:main"

# Individual tools
qsp-enrich-csv = "maple.cli.enrich:main"
qsp-export-model = "maple.cli.export_model:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["maple*"]

[tool.setuptools.package-data]
"maple" = [
    "templates/**/*",
    "prompts/**/*",
]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
]
markers = [
    "integration: marks tests that hit external APIs (deselect with '-m \"not integration\"')",
]
# Ignore Pydantic models that start with "Test" (not test classes)
filterwarnings = [
    "ignore::pytest.PytestCollectionWarning:maple.core.pydantic_models",
]
